feat: Todos os MainSave e relações com todas as classes

+++ src/main/java/br/edu/ifpb/es/daw/entities/Presenca.java
@@ -13,7 +13,7 @@ public class Presenca {
 
     }
 
-    public Presenca(Aluno aluno, Aula aula, Integer id, StatusPresenca status) {
+    public Presenca(Aluno aluno, Aula aula, Long id, StatusPresenca status) {
         this.aluno = aluno;
         this.aula = aula;
         this.id = id;
@@ -22,7 +22,7 @@ public class Presenca {
 
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
-    private Integer id;
+    private Long id;
 
 
     @ManyToOne(cascade = CascadeType.PERSIST)
@@ -63,11 +63,11 @@ public class Presenca {
         this.status = status;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
